Move some bits of code to `oonipi` module
Also, re-run mubench/hash64_shingles to check for possible regressions.
