+++ af/shovel/canning.py
@@ -18,9 +18,10 @@ import traceback
 from base64 import b64encode
 from contextlib import closing, contextmanager
 from hashlib import sha1
-from subprocess import Popen, PIPE
 from zlib import crc32 # btw, binascii.crc32 is four times slower than zlib.crc32
 
+from oonipl.cli import dirname, isomidnight
+from oonipl.popen import ScopedPopen, PIPE
 
 class CannerError(RuntimeError):
     pass
@@ -62,17 +63,6 @@ class Future(object):
         else:
             return self.retval
 
-
-@contextmanager
-def ScopedPopen(*args, **kwargs):
-    proc = Popen(*args, **kwargs)
-    try:
-        yield proc
-    finally:
-        try:
-            proc.kill()
-        except Exception:
-            pass
 
 
 class NopTeeFd(object):
@@ -457,26 +447,6 @@ def canning(input_root, output_root, bucket):
         os.link(fdindex.name, index_fpath)
     os.chmod(index_fpath, 0444)
     os.chmod(output_dir, 0555) # done!
-
-
-def isodatetime(s):
-    # Reverse of `datetime.isoformat()` besides microseconds & timezones
-    return datetime.datetime.strptime(s, '%Y-%m-%dT%H:%M:%S')
-
-
-def isomidnight(s):
-    s = isodatetime(s)
-    if s.hour == s.minute == s.second == s.microsecond == 0:
-        return s
-    raise ValueError('isodatetime is not midnight-aligned', s)
-
-
-def dirname(s):
-    if not os.path.isdir(s):
-        raise ValueError('Not a directory', s)
-    if s[-1] == '/':
-        raise ValueError('Bogus trailing slash', s)
-    return s
 
 
 def parse_args():
