+++ af/shovel/oonipl/__init__.py
@@ -1,0 +1,2 @@
+# perl code of OONI written in python
+# ... because it's possible to write perl code in any language!
