+++ af/shovel/oonipl/tor_log.py
@@ -1,4 +1,5 @@
-#!/usr/bin/env python2
+#!/usr/bin/env python2.7
+# -*- coding: utf-8 -*-
 
 import re
